requirements.txt
allure-pytest==2.10.0
allure-python-commons==2.10.0
assertpy~=1.1
axe-selenium-python~=2.1.6
clients~=1.3
config~=0.5.1
pipenv==2022.9.8
pytest
pytest-reportportal~=5.1.2
pytest-xdist
selenium==4.4.3
virtualenv==20.16.5
virtualenv-clone==0.5.7
webdriver-manager==3.8.3
appium-python-client>=2.6.1
requests==2.28.1
